BACKEND/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-Admin
python-dotenv